Doc updated.

+++ forms.py
@@ -1,6 +1,6 @@
 from flask_wtf import FlaskForm
 from flask_login import current_user
-from wtforms import StringField,PasswordField,SubmitField,RadioField,DateField,BooleanField, DecimalField
+from wtforms import StringField,PasswordField,SubmitField,RadioField,DateField,BooleanField, IntegerField
 from wtforms.validators import DataRequired, Length, Email,EqualTo, ValidationError
 from userdb import username_check, mail_check
 from books import check_tpage
@@ -35,16 +35,16 @@ class loginForm(FlaskForm):
 class tvForm(FlaskForm):
     title= StringField('Title',validators=[DataRequired()])
     language= StringField('Language',validators=[DataRequired()])
-    year= DecimalField('Year',validators=[DataRequired()])
-    season= DecimalField('Season',validators=[DataRequired()])
+    year= IntegerField('Year',validators=[DataRequired()])
+    season= IntegerField('Season',validators=[DataRequired()])
     genre= StringField('Genre',validators=[DataRequired()])
     channel= StringField('Channel',validators=[DataRequired()])
     submit = SubmitField('Add Tv Series')
 
 class episodeForm(FlaskForm):
     title= StringField('Title',validators=[DataRequired()])
-    season= DecimalField('Language',validators=[DataRequired()])
-    episode= DecimalField('Year',validators=[DataRequired()])
+    season= IntegerField('Language',validators=[DataRequired()])
+    episode= IntegerField('Year',validators=[DataRequired()])
     submit = SubmitField('Add Episode')
 
 
@@ -52,8 +52,8 @@ class episodeForm(FlaskForm):
 class bookForm(FlaskForm):
     name= StringField('Title',validators=[DataRequired()])
     writer = StringField('Author',validators=[DataRequired()])
-    year_pub = DecimalField('Year of Publication',validators=[DataRequired()])
-    tpage = DecimalField('Total Page',validators=[DataRequired()])
+    year_pub = IntegerField('Year of Publication',validators=[DataRequired()])
+    tpage = IntegerField('Total Page',validators=[DataRequired()])
     publisher = StringField('Publisher',validators=[DataRequired()])
     language = StringField('Language',validators=[DataRequired()])
     genre = StringField('Genre',validators=[DataRequired()])
